reworked repos and services to reduce code scatter
