--- app/src/main/java/com/yalta/services/LogoutService.kt
@@ -1,18 +1,0 @@
-package com.yalta.services
-
-import com.yalta.repositories.LogoutRepo
-import com.yalta.repositories.process
-import javax.inject.Inject
-
-class LogoutService @Inject constructor(private val repo: LogoutRepo) {
-    suspend fun logout(): Boolean {
-        return process(
-            { repo.logout() },
-            {
-                SessionService.discardSession()
-                true
-            },
-            { false }
-        )
-    }
-}
