+++ app/src/main/java/com/yalta/services/PasswordService.kt
@@ -5,7 +5,9 @@ import com.yalta.repositories.Reason
 import com.yalta.repositories.process
 import java.util.*
 import javax.inject.Inject
+import javax.inject.Singleton
 
+@Singleton
 class PasswordService @Inject constructor(private val repo: PasswordRepo) {
     suspend fun changePassword(newPassword: String): Optional<Boolean> {
         return process(
