--- app/src/main/java/com/yalta/services/UserService.kt
@@ -1,16 +1,0 @@
-package com.yalta.services
-
-import com.yalta.repositories.UserRepo
-import com.yalta.repositories.process
-import common.User
-import javax.inject.Inject
-
-class UserService @Inject constructor(private val repo: UserRepo) {
-    suspend fun getUser(): User? {
-        return process(
-            { repo.getUser() },
-            { it.user },
-            { null }
-        )
-    }
-}
